refactor(live-sessions): extract seat-from-screenshot logic into hook
Move state machine (step/sourceApp/rows/isApplying), tRPC query+mutation,
row processing, and query invalidation out of seat-from-screenshot-sheet.tsx
into a new use-seat-from-screenshot hook. Pure helpers (normalizeName,
buildRow, applyRow, applyRowAction, fileToBase64, computeRowAction,
computeRowWarning, ACCEPTED_TYPES, SOURCE_APP_ENTRIES) and types
(SessionParam, Step, RowAction, ReviewRow, PlayerOption) move to
seat-screenshot utils module.

Invalidation uses invalidateTargets from the common optimistic-update helper.
Component shrinks from 923 to ~460 lines, presenting state + handlers from
the hook. ReviewRowItem and SeatCombobox remain co-located as
presentational sub-components.
